Refactoring Container StackView for filters RefreshControl for pull to refresh Remove ActivityIndicator Dependency injection by method for cell Global URL for signUpLabel (remove attribute)
