+++ Flickr/Flickr/Views/Labels/SignUpLabel.swift
@@ -10,6 +10,7 @@ import SafariServices
 
 class SignUpLabel: UILabel {
     var delegate: UIViewController?
+    var url = URL(string: "https://identity.flickr.com/sign-up")!
     
     func setup() {
         let fontSize: CGFloat = 14
@@ -17,10 +18,7 @@ class SignUpLabel: UILabel {
                                                    attributes: [.font: UIFont.systemFont(ofSize: fontSize, weight: .light),.foregroundColor: UIColor.secondaryLabel, .kern: 0.5])
         let attrbs: [NSAttributedString.Key: Any] = [
             .foregroundColor: UIColor.label,
-            .font: UIFont.systemFont(ofSize: fontSize, weight: .semibold),
-            .attachment: URL(string: "https://identity.flickr.com/sign-up")!,
-            .underlineStyle: 0,
-            .underlineColor: UIColor.clear]
+            .font: UIFont.systemFont(ofSize: fontSize, weight: .semibold)]
         textString.addAttributes(attrbs, range: NSRange(location: 23, length: 8))
         attributedText = textString
         isUserInteractionEnabled = true
@@ -29,8 +27,6 @@ class SignUpLabel: UILabel {
     }
     
     @objc private func showSignUp() {
-        guard let characterCount = self.text?.count else { return }
-        guard let url = self.attributedText?.attribute(.attachment, at: characterCount - 1, effectiveRange: nil) as? URL else { return } // characterCount - 1, because we want to get last index in self.text
         let signUpVC = SignUpViewController()
         signUpVC.url = url
         signUpVC.title = "Sign Up"
@@ -39,7 +35,6 @@ class SignUpLabel: UILabel {
         let vc = UINavigationController(rootViewController: signUpVC)
         vc.modalPresentationStyle = .popover
         delegate?.show(vc, sender: nil)
-        
     }
     
     @objc private func closeScreen() {
